+++ src/tests/test_interface.py
@@ -71,12 +71,29 @@ def test_override():
     def bar(self):
       return 42
 
-  with pytest.raises(ImplementationError):
+  with pytest.raises(ImplementationError) as excinfo:
+    @implements(IFoo)
+    class Foo(object):
+      @override
+      def bar(self):
+        return 42
+      @override
+      def bars(self):
+        return 42
+  assert excinfo.value.interfaces == []
+  assert excinfo.value.errors == ["'bars' does not override a method of any of the implemented interfaces."]
+
+  with pytest.raises(ImplementationError) as excinfo:
     @implements(IFoo)
     class Foo(object):
       @override
       def bars(self):
         return 42
+  assert excinfo.value.interfaces == [IFoo]
+  assert excinfo.value.errors == [
+    "'bars' does not override a method of any of the implemented interfaces.",
+    "missing method: bar()"
+  ]
 
 
 def test_interface_constructor():
@@ -133,7 +150,7 @@ def test_default():
   assert Bar() == '52'
 
 
-class test_staticmethod_override():
+def test_staticmethod_override():
 
   class IFoo(Interface):
     @staticmethod
@@ -155,7 +172,7 @@ class test_staticmethod_override():
   #                    does not work.
 
 
-class test_attr_default():
+def test_attr_default():
 
   class IFoo(Interface):
     x = attr(int, 24)
@@ -169,7 +186,7 @@ class test_attr_default():
   assert Bar().x == 24
 
 
-class test_staticattr_default():
+def test_staticattr_default():
 
   class IFoo(Interface):
     x = staticattr(24)
@@ -182,3 +199,156 @@ class test_staticattr_default():
   assert Bar.x == 24
   assert Bar().x == 24
   assert 'x' not in vars(Bar())
+
+
+def test_override_works_on_staticmethod():
+  """
+  In Python 2.7 you cannot set a member on an instance of the [[staticmethod]]
+  class. This means a previous implementation of [[override]] resulted in an
+  [[AttributeError]].
+  """
+
+  class IFoo(Interface):
+    @staticmethod
+    def a_static_method():
+      pass
+
+  @implements(IFoo)
+  class Bar(object):
+    @override
+    @staticmethod
+    def a_static_method():
+      pass
+
+
+def test_property_missing():
+
+  class IFoo(Interface):
+    @property
+    def foo(self):
+      pass
+
+  with pytest.raises(ImplementationError) as excinfo:
+    @implements(IFoo)
+    class Bar(object):
+      pass
+  assert excinfo.value.interfaces == [IFoo]
+  assert excinfo.value.errors == ["missing property: foo"]
+
+
+def test_property_wrongtype():
+
+  class IFoo(Interface):
+    @property
+    def foo(self):
+      pass
+
+  with pytest.raises(ImplementationError) as excinfo:
+    @implements(IFoo)
+    class Bar(object):
+      def foo(self):
+        pass
+  assert excinfo.value.interfaces == [IFoo]
+  assert excinfo.value.errors == ["expected property, got instancemethod: foo"]
+
+
+def test_property_ok():
+
+  class IFoo(Interface):
+    @property
+    def foo(self):
+      pass
+
+  @implements(IFoo)
+  class Bar(object):
+    @property
+    def foo(self):
+      pass
+
+    # This works because this is not necessarily a semantic that is
+    # incompatible with the interface.
+    @foo.setter
+    def foo(self, value):
+      pass
+
+
+def test_property_missing_setter():
+
+  class IFoo(Interface):
+    @property
+    def foo(self):
+      pass
+
+    @foo.setter
+    def foo(self, value):
+      pass
+
+  with pytest.raises(ImplementationError) as excinfo:
+    @implements(IFoo)
+    class Bar(object):
+      @property
+      def foo(self):
+        pass
+  assert excinfo.value.interfaces == [IFoo]
+  assert excinfo.value.errors == ["property foo: missing setter"]
+
+
+def test_property_final():
+
+  class AError(Exception):
+    pass
+
+  class BError(Exception):
+    pass
+
+  class IFoo(Interface):
+    @property
+    def foo(self):
+      pass
+
+    @foo.setter
+    @final
+    def foo(self, value):
+      raise AError
+
+  @implements(IFoo)
+  class Bar(object):
+    @property
+    def foo(self):
+      raise BError
+
+  with pytest.raises(AError):
+    Bar().foo = 42
+
+  with pytest.raises(BError):
+    Bar().foo
+
+  with pytest.raises(ImplementationError) as excinfo:
+    @implements(IFoo)
+    class Bar(object):
+      @property
+      def foo(self):
+        pass
+
+      @foo.setter
+      def foo(self, value):
+        pass
+  assert excinfo.value.interfaces == [IFoo]
+  assert excinfo.value.errors == ["property foo: setter must not be implemented"]
+
+
+def test_property_wrong_decoration():
+
+  # TODO(nrosenstein): This should actually raise an error
+
+  class IFoo(Interface):
+    @property
+    def foo(self):
+      pass
+
+    @final
+    @foo.setter
+    def foo(self, value):
+      raise AError
+
+  assert 'foo' not in IFoo